[RED] create tests for useQuizBuilder hook
This hook will be the source of truth for the quiz builder tool. 99% of the state will be managed here. The hook will be capable of changing question type, updating questions, and deleting them. The hook will also be able to validate the entire quiz before being posted to the server